move authenticator tests to dedicated file
